+++ beemapi/__init__.py
@@ -3,7 +3,6 @@ from .version import version as __version__
 __all__ = [
     "noderpc",
     "exceptions",
-    "websocket",
     "rpcutils",
     "graphenerpc",
     "node",